extracts context_menu helper method into own file
